Prevent browser default unload behavior
When returning a non-`null` and non-`undefined` value from an `onbeforeunload` handler, causes some browsers to show a generic warning dialog.

Before this change, the handler would determine that there are no unsaved changes if there are none, but it would still return a value that signals the browser to show the default dialog.

This change achieves that no dialog is shown if there are no unsaved changes or if we're in a view state where this behavior is undesirable.

Fixes #1633